+++ src/altrep-rep.h
@@ -1,18 +1,9 @@
 #ifndef VCTRS_ALTREP_REP_H
 #define VCTRS_ALTREP_REP_H
 
-#include "altrep.h"
-
 bool vec_is_altrep_vctrs_compact_rep(SEXP x);
 
 SEXP new_altrep_vctrs_compact_rep_int(int value, R_xlen_t size);
 SEXP new_altrep_vctrs_compact_rep_dbl(double value, R_xlen_t size);
-
-#if (R_VERSION >= R_Version(3, 5, 0))
-
-R_altrep_class_t altrep_vctrs_compact_rep_int_class;
-R_altrep_class_t altrep_vctrs_compact_rep_dbl_class;
-
-#endif
 
 #endif
